+++ sourceCode/dynamics/RigidBodyDyn_base.cpp
@@ -50,7 +50,7 @@ C7Vector CRigidBodyDyn_base::getShapeFrameTransformation()
     return(C7Vector::identityTransformation);
 }
 
-void CRigidBodyDyn_base::reportVelocityToShape(float simulationTime)
+void CRigidBodyDyn_base::reportVelocityToShape(double simulationTime)
 {
 }
 
@@ -58,7 +58,7 @@ void CRigidBodyDyn_base::handleAdditionalForcesAndTorques()
 {
 }
 
-void CRigidBodyDyn_base::handleKinematicBody_step(float t,float cumulatedTimeStep)
+void CRigidBodyDyn_base::handleKinematicBody_step(double t,double cumulatedTimeStep)
 {
 }
 
@@ -86,7 +86,7 @@ CXShape* CRigidBodyDyn_base::getShape() const
     return(_shape);
 }
 
-void CRigidBodyDyn_base::reportConfigurationToShape(float simulationTime)
+void CRigidBodyDyn_base::reportConfigurationToShape(double simulationTime)
 {
     if (!_isStatic)
     { // dynamic
@@ -105,7 +105,7 @@ void CRigidBodyDyn_base::reportConfigurationToShape(float simulationTime)
     }
 }
 
-void CRigidBodyDyn_base::handleKinematicBody_init(float dt)
+void CRigidBodyDyn_base::handleKinematicBody_init(double dt)
 {
     if (_isStatic)
     {
@@ -116,21 +116,21 @@ void CRigidBodyDyn_base::handleKinematicBody_init(float dt)
         C7Vector cumulTrp1;
         _simGetObjectCumulativeTransformation(_shape,cumulTrp1.X.data,cumulTrp1.Q.data,true);
         _bodyEnd_kinematicBody=cumulTrp1*aax;
-        _applyBodyToShapeTransf_kinematicBody=( ((_bodyStart_kinematicBody.X-_bodyEnd_kinematicBody.X).getLength()>0.00001f)||(_bodyStart_kinematicBody.Q.getAngleBetweenQuaternions(_bodyEnd_kinematicBody.Q)>0.05f*degToRad) );
+        _applyBodyToShapeTransf_kinematicBody=( ((_bodyStart_kinematicBody.X-_bodyEnd_kinematicBody.X).getLength()>0.00001)||(_bodyStart_kinematicBody.Q.getAngleBetweenQuaternions(_bodyEnd_kinematicBody.Q)>0.05*degToRad) );
         
         C3Vector dx;
         _simGetInitialDynamicVelocity(_shape,dx.data);
         dx*=dt;
         C3Vector dxa;
         _simGetInitialDynamicAngVelocity(_shape,dxa.data);
-        if (dx.getLength()+dxa.getLength()>0.0f)
+        if (dx.getLength()+dxa.getLength()>0.0)
         {
             _simSetInitialDynamicVelocity(_shape,C3Vector::zeroVector.data); // important to reset it!
             _simSetInitialDynamicAngVelocity(_shape,C3Vector::zeroVector.data); // important to reset it!
             C7Vector tr;
             _simGetObjectCumulativeTransformation(_shape,tr.X.data,tr.Q.data,true);
             tr.X+=dx;
-            if (dxa.getLength()>0.0f)
+            if (dxa.getLength()>0.0)
             { // following new since 25/03/2013
                 C4Vector q(dxa(0),dxa(1),dxa(2));
                 C4Vector angleAndAxis(q.getAngleAndAxis());
